+++ src/main/java/com/future/entity/commission/FuCommissionCustomer.java
@@ -42,12 +42,12 @@ public class FuCommissionCustomer {
     /**
      * 原订单状态
      */
-    private Byte sourceOrderState;
+    private int sourceOrderState;
 
     /**
      * 原订单操作类型
      */
-    private Byte sourceOrderOper;
+    private int sourceOrderOper;
 
     /**
      * 源金额
@@ -57,7 +57,7 @@ public class FuCommissionCustomer {
     /**
      * 佣金类型（0 信号源返佣，1 多级返佣，2 社区返佣，3 其他）
      */
-    private Byte commissionType;
+    private int commissionType;
 
     /**
      * 收佣用户ID
@@ -67,12 +67,12 @@ public class FuCommissionCustomer {
     /**
      * 收佣用户类型
      */
-    private Byte commissionUserType;
+    private int commissionUserType;
 
     /**
      * 收佣用户等级
      */
-    private Byte commissionUserLevel;
+    private int commissionUserLevel;
 
     /**
      * 收佣账户ID
@@ -97,7 +97,7 @@ public class FuCommissionCustomer {
     /**
      * 返佣状态（0 成功，1失败，2 执行中）
      */
-    private Byte commissionState;
+    private int commissionState;
 
     /**
      * 
@@ -215,7 +215,7 @@ public class FuCommissionCustomer {
      * 原订单状态
      * @return source_order_state 原订单状态
      */
-    public Byte getSourceOrderState() {
+    public int getSourceOrderState() {
         return sourceOrderState;
     }
 
@@ -223,7 +223,7 @@ public class FuCommissionCustomer {
      * 原订单状态
      * @param sourceOrderState 原订单状态
      */
-    public void setSourceOrderState(Byte sourceOrderState) {
+    public void setSourceOrderState(int sourceOrderState) {
         this.sourceOrderState = sourceOrderState;
     }
 
@@ -231,7 +231,7 @@ public class FuCommissionCustomer {
      * 原订单操作类型
      * @return source_order_oper 原订单操作类型
      */
-    public Byte getSourceOrderOper() {
+    public int getSourceOrderOper() {
         return sourceOrderOper;
     }
 
@@ -239,7 +239,7 @@ public class FuCommissionCustomer {
      * 原订单操作类型
      * @param sourceOrderOper 原订单操作类型
      */
-    public void setSourceOrderOper(Byte sourceOrderOper) {
+    public void setSourceOrderOper(int sourceOrderOper) {
         this.sourceOrderOper = sourceOrderOper;
     }
 
@@ -263,7 +263,7 @@ public class FuCommissionCustomer {
      * 佣金类型（0 信号源返佣，1 多级返佣，2 社区返佣，3 其他）
      * @return commission_type 佣金类型（0 信号源返佣，1 多级返佣，2 社区返佣，3 其他）
      */
-    public Byte getCommissionType() {
+    public int getCommissionType() {
         return commissionType;
     }
 
@@ -271,7 +271,7 @@ public class FuCommissionCustomer {
      * 佣金类型（0 信号源返佣，1 多级返佣，2 社区返佣，3 其他）
      * @param commissionType 佣金类型（0 信号源返佣，1 多级返佣，2 社区返佣，3 其他）
      */
-    public void setCommissionType(Byte commissionType) {
+    public void setCommissionType(int commissionType) {
         this.commissionType = commissionType;
     }
 
@@ -295,7 +295,7 @@ public class FuCommissionCustomer {
      * 收佣用户类型
      * @return commission_user_type 收佣用户类型
      */
-    public Byte getCommissionUserType() {
+    public int getCommissionUserType() {
         return commissionUserType;
     }
 
@@ -303,7 +303,7 @@ public class FuCommissionCustomer {
      * 收佣用户类型
      * @param commissionUserType 收佣用户类型
      */
-    public void setCommissionUserType(Byte commissionUserType) {
+    public void setCommissionUserType(int commissionUserType) {
         this.commissionUserType = commissionUserType;
     }
 
@@ -311,7 +311,7 @@ public class FuCommissionCustomer {
      * 收佣用户等级
      * @return commission_user_level 收佣用户等级
      */
-    public Byte getCommissionUserLevel() {
+    public int getCommissionUserLevel() {
         return commissionUserLevel;
     }
 
@@ -319,7 +319,7 @@ public class FuCommissionCustomer {
      * 收佣用户等级
      * @param commissionUserLevel 收佣用户等级
      */
-    public void setCommissionUserLevel(Byte commissionUserLevel) {
+    public void setCommissionUserLevel(int commissionUserLevel) {
         this.commissionUserLevel = commissionUserLevel;
     }
 
@@ -391,7 +391,7 @@ public class FuCommissionCustomer {
      * 返佣状态（0 成功，1失败，2 执行中）
      * @return commission_state 返佣状态（0 成功，1失败，2 执行中）
      */
-    public Byte getCommissionState() {
+    public int getCommissionState() {
         return commissionState;
     }
 
@@ -399,7 +399,7 @@ public class FuCommissionCustomer {
      * 返佣状态（0 成功，1失败，2 执行中）
      * @param commissionState 返佣状态（0 成功，1失败，2 执行中）
      */
-    public void setCommissionState(Byte commissionState) {
+    public void setCommissionState(int commissionState) {
         this.commissionState = commissionState;
     }
 }
